Add initial support for Unified Memory backend (#460)
* Add initial support for Unified Memory

* Add MGPU support for transform points

* Cleanup

* Add guard for CUDA version

* Add cudamanaged support for JOffsetsFromJIdx and PointsInGrid

* Fixes for signed/unsigned warnings

* Rename for consistency

* Remove hardcoded launch bounds

* Internalize numThreads to forEach kernels

* Rename macro to explicitly call out backends

* Clean up signed/unsigned indexing and reduce default block size

* Clean up macro

* Format code

* Remove unnecessary includes

* Address some review comments

* Fix helper function and formatting

* Encapsulate allocation logic

+++ fvdb/src/detail/ops/ReadFromDense.cu
@@ -56,7 +56,7 @@ readFromDenseCPU(const GridBatchImpl::Accessor<nanovdb::ValueOnIndex> &gridHandl
                  const torch::TensorAccessor<int32_t, 2> denseOrigins,
                  torch::TensorAccessor<ScalarType, 2> outSparseTensor,
                  bool isContiguous) {
-    for (size_t bi = 0; bi < gridHandle.batchSize(); bi += 1) {
+    for (int64_t bi = 0; bi < gridHandle.batchSize(); bi += 1) {
         const nanovdb::OnIndexGrid *grid = gridHandle.grid(bi);
         const nanovdb::Coord denseOrigin(
             denseOrigins[bi][0], denseOrigins[bi][1], denseOrigins[bi][2]);
